refactor(services): extract shared emitClinicalEvent from clinical-data and clinical-notes
The `emitClinicalEvent` function was byte-identical in
`services/clinical-data/src/events.ts` and `services/clinical-notes/src/events.ts`
after PR #806. Move the single source of truth into `@carebridge/outbox`,
which already groups related clinical-event primitives (writeOutboxEntry,
readPendingBatch, markProcessed, etc.). The service-level events.ts files
become thin re-export shims so the many existing call sites and their tests
keep working unchanged.

- Add `packages/outbox/src/emit.ts` with `emitClinicalEvent`.
- Re-export it from `packages/outbox/src/index.ts`.
- Add `bullmq` and `@carebridge/redis-config` to outbox dependencies.
- Replace the duplicate `services/{clinical-data,clinical-notes}/src/events.ts`
  with two-line re-exports from `@carebridge/outbox`.
- Move the unit test to `packages/outbox/src/__tests__/emit.test.ts` —
  it tests the shared implementation directly.

Closes #817
